+++ src/models/StyledNativeComponent.js
@@ -7,6 +7,7 @@ import generateDisplayName from '../utils/generateDisplayName'
 import isStyledComponent from '../utils/isStyledComponent'
 import isTag from '../utils/isTag'
 import hasInInheritanceChain from '../utils/hasInInheritanceChain'
+import once from '../utils/once'
 import { CHANNEL_NEXT, contextShape } from './ThemeProvider'
 
 import type { Theme } from './ThemeProvider'
@@ -15,6 +16,16 @@ import type { RuleSet, Target } from '../types'
 type State = {
   theme?: ?Theme,
   generatedStyles: any,
+}
+
+let warnExtendDeprecated
+if (process.env.NODE_ENV !== 'production') {
+  warnExtendDeprecated = once(() => {
+    // eslint-disable-next-line no-console
+    console.error(
+      'Warning: The "extend" API will be removed in the upcoming v4.0 release. Use styled(StyledComponent) instead. You can find more information here: https://github.com/styled-components/styled-components/issues/1546'
+    )
+  })
 }
 
 // $FlowFixMe
@@ -225,6 +236,8 @@ export default (constructWithOptions: Function, InlineStyle: Function) => {
           rules: newRules,
           ParentComponent: StyledNativeComponent,
         }
+
+        warnExtendDeprecated()
 
         return constructWithOptions(
           createStyledNativeComponent,